requirements.txt
uvicorn==0.24.0.post1
bcrypt==4.1.1
cryptography==41.0.7
fastapi==0.104.1
passlib==1.7.4
pydantic==2.5.2
PyJWT==2.8.0
python-multipart==0.0.6
SQLAlchemy==2.0.23
starlette==0.27.0
uvicorn==0.24.0.post1
flake8==6.1.0